+++ OSBarcodeLib/Scanner/OSBARCScannerBehaviour.swift
@@ -10,7 +10,7 @@ final class OSBARCScannerBehaviour: OSBARCCoordinatable, OSBARCScannerProtocol {
     /// The publisher's cancellable instance collector.
     private var cancellables: Set<AnyCancellable> = []
     
-    func startScanning(with instructionsText: String, and buttonText: String?, _ completion: @escaping (String) -> Void) {
+    func startScanning(with instructionsText: String, _ buttonText: String?, and cameraModel: OSBARCCameraModel, _ completion: @escaping (String) -> Void) {
         $scanResult
             .dropFirst()    // drops the first value - the empty string
             .first()        // only publishes the first barcode value found
@@ -30,7 +30,8 @@ final class OSBARCScannerBehaviour: OSBARCCoordinatable, OSBARCScannerProtocol {
         )
 
         // Get the default camera for capturing videos. This object will allows us to fetch the `hasTorch` method.
-        let captureDevice = AVCaptureDevice.default(.builtInWideAngleCamera, for: .video, position: .back)
+        let cameraToUse = AVCaptureDevice.Position.map(cameraModel)
+        let captureDevice = AVCaptureDevice.default(.builtInWideAngleCamera, for: .video, position: cameraToUse)
         let cameraHasTorch = captureDevice?.hasTorch ?? false
         
         let buttonText = buttonText ?? ""   // not having the button enabled is translated into having an empty text.